requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==20.7
aiosqlite==0.20.0
python-dotenv==1.0.1
websockets==12.0
